+++ lib/kuality_ole/base_objects/info/marc_data_line.rb
@@ -17,7 +17,7 @@ class MarcDataLine < InfoObject
 
   include KualityOle::Helpers
 
-  attr_accessor :tag,:indicator_1,:indicator_2,:subfield_code,:subfield,:value
+  attr_accessor :tag,:indicator_1,:indicator_2,:subfield_codes,:subfield,:values
 
   alias :ind_1 :indicator_1
   alias :ind_1= :indicator_1=
@@ -28,29 +28,59 @@ class MarcDataLine < InfoObject
   #   :tag              String        The MARC field tag
   #   :indicator_1      String        The first subfield indicator.
   #   :indicator_2      String        The second subfield indicator.
-  #   :subfield_code    String        The MARC subfield code/subfield delimiter.
-  #   :value            String        The actual value of the field, without delimiter.
+  #   :subfield_codes   Array         An array of MARC subfield codes (delimiters).
+  #   :values           String        An array of MARC subfield values (without delimiter).
+  #                                  
+  # Usage:
+  #   -- Single Subfield --
+  #
+  #   MarcDataLine.new(:tag => '100',:subfield_codes => ['$a'],:values => ['Arthur Authorsen'])
+  #
+  #   -- Multiple Subfields --
+  #   MarcDataLine.new(:tag => '245',
+  #     :subfield_codess => ['$a','$b'],
+  #     :values => ['Title of Book','Subtitle of book']
+  #   )
   #
   # @note In the OLE Library System, MARC subfields are delimited with a
   #   vertical bar '|' instead of the standard dollar sign '$', and the
-  #   subfield code is usually a capital letter.
+  #   subfield code is required to be a lowercase letter.
   #   These substitutions will be made automatically.
   #
   def initialize(opts={})
     defaults = {
-        :tag          => '245',
-        :subfield     => '|a',
-        :indicator_1  => '#',
-        :indicator_2  => '#',
-        :value        => random_letters(pick_range(6..10)).capitalize
+        :tag              => '245',
+        :indicator_1      => '#',
+        :indicator_2      => '#',
+        :subfield_codes   => ['|a'],
+        :values           => [random_letters(pick_range(6..10)).capitalize]
     }
     options = defaults.merge(opts)
 
     @tag              = options[:tag]
-    @subfield_code    = options[:subfield].gsub(/(?<=^)\$/,'|').downcase
+    @subfield_codes   = options[:subfield_codes].collect {|sfc| sfc.gsub(/(?<=^)\$/,'').gsub(/^(?!\|)/,'|').downcase}
     @indicator_1      = options[:indicator_1]
     @indicator_2      = options[:indicator_2]
-    @value            = options[:value]
-    @subfield         = "#{@subfield_code} #{value}"
+    @values           = options[:values]
+    @subfield         = options.has_key?(:subfield) ? options[:subfield] : [@subfield_codes,@values].transpose.flatten.join(' ')
+  end
+
+  class << self
+    # Create a new instance of MarcDataLine by passing in a MARC::DataField from Ruby-Marc.
+    #
+    # Params:
+    #   obj     Object      An instance of a MARC::DataField from the Ruby-Marc gem.
+    #
+    def from_field(obj)
+      opts = {
+        :tag              => obj.tag,
+        :indicator_1      => obj.indicator1,
+        :indicator_2      => obj.indicator2,
+        :subfield_codes   => obj.subfields.collect {|sf| sf.code},
+        :values           => obj.subfields.collect {|sf| sf.value}
+      }
+      self.new(opts)
+    end
+    alias_method(:new_from_field,:from_field)
   end
 end