+++ spec/data/marc_records_spec.rb
@@ -15,28 +15,3 @@
 require 'rspec'
 require 'spec_helper'
 
-describe 'A Marc Data Line' do
-
-  let(:marc_data_line)       {MarcDataLine.new}
-
-  it 'has a default tag value' do
-    expect(marc_data_line.tag).to eq('245')
-  end
-
-  it 'has default indicator values' do
-    expect(marc_data_line.indicator_1).to eq('#')
-    expect(marc_data_line.indicator_2).to eq('#')
-  end
-
-  it 'has a default subfield code' do
-    expect(marc_data_line.subfield_code).to eq('|a')
-  end
-
-  it 'has a full subfield value' do
-    expect(marc_data_line.subfield).to match(/^\|a [\w]+/)
-  end
-
-  it 'has an actual data value' do
-    expect(marc_data_line.value).to match(/^[\w]+$/)
-  end
-end